feature: working on regression analysis
