+++ src/utils/pipeline.py
@@ -21,8 +21,6 @@ import subprocess
 import pandas
 import joblib
 
-from sklearn.utils import resample
-
 
 def parse_args(message: str = "",
                return_parser: bool = False
@@ -120,64 +118,3 @@ def create_summary(dataframe: pandas.DataFrame) -> dict:
                 round(dataframe['survived'].value_counts(normalize=True, dropna=False), 3)),
             'features': dataframe.columns.tolist()
             }
-
-
-def resample_by_category(target: str,
-                         x_train: pandas.DataFrame,
-                         up_or_down: str = 'up',
-                         resampling_perc: float = 1.0) -> pandas.DataFrame:
-    """
-    Method to resample the training data
-
-    Args:
-        target (str)
-        x_train (pandas.DataFrame)
-        up_or_down (str): By default = 'up'
-        resampling_perc (float): By default = 1.0
-
-    Returns:
-        x_train (pandas.DataFrame)
-    """
-    majority_label = x_train[target].mode()[0]
-    majority_data = x_train[x_train[target] == majority_label]
-    minority_data = x_train[x_train[target] != majority_label]
-
-    if resampling_perc > 2.0 or resampling_perc <= 0.01:
-        raise Exception(
-            f'Invalid value {resampling_perc} for parameter resampling_perc, values must be between 0.01 and 2.00')
-
-    up_or_down = up_or_down.lower().strip()
-
-    if up_or_down == 'up':
-
-        print(f'\t-> Upsampling minority class\n' +
-              f'\t\t Minority class will be resampled to {majority_data.shape[0]} number of rows')
-
-        data2resample = minority_data
-        n_samples = int(round(majority_data.shape[0] * resampling_perc, 0))
-        data2join = majority_data
-
-    elif up_or_down == 'down':
-
-        print(f'\t-> Downsampling majority class\n' +
-              f'\t\t Majority class will be resampled to {minority_data.shape[0]} number of rows')
-
-        data2resample = majority_data
-        n_samples = int(round(minority_data.shape[0] * resampling_perc, 0))
-        data2join = minority_data
-
-    else:
-        raise Exception(f'Invalid value {up_or_down} for variable up_or_down. Valid values are ["up","down"]')
-
-    resampled_data = resample(
-        data2resample,
-        replace=True,
-        n_samples=n_samples,
-        random_state=1234
-    )
-
-    x_train = pandas.concat([resampled_data, data2join])
-
-    print(f'\n\t\tNew proportion of targets: {x_train[target].value_counts(normalize=True).to_dict()}\n')
-
-    return x_train
